added bootstrap_suites option to default environment in quickstart-vbox database clean paella-init-nfsinstaller a bit, since it used to say "skipping install" after a successful install ask for confirmation before deleting a machine_type script added warnings to runlog, so we can start passing lists to runlog, 	instead of strings, and locate places where failure_suppressed 	is passed as True (probably places where it shouldn't be) log the keys that are added to apt during the install created new kernel helper class to help install both grub and kernel 	not a great solution, but the class helps replace a chunk of 	even uglier code in util.preinst, we create a resolv.conf file from the host

+++ src/paella/kde/machine/viewbrowser.py
@@ -108,7 +108,9 @@ class MachineTypeView(ViewBrowser):
         elif context == 'Variables':
             self.doc.mtype.edit_variables()
         elif context == 'Scripts':
-            self.doc.mtype.delete_script(ident)
+            ans = KMessageBox.questionYesNo(self, "really delete this script?")
+            if ans == KMessageBox.Yes:
+                self.doc.mtype.delete_script(ident)
         elif context == 'Modules':
             msg = 'Deleting modules is not supported.'
             KMessageBox.information(self, msg)
